architectures in own file per type, consistent naming
