+++ examples/experiments/ptb-joint-pos-dep/defaults.cfg
@@ -39,24 +39,24 @@ factory = "tagger"
 factory = "parser"
 
 [nlp.pipeline.tagger.model]
-@architectures = "tagger_model.v1"
+@architectures = "spacy.TaggerModel.v1"
 
 [nlp.pipeline.tagger.model.tok2vec]
-@architectures = "tok2vec_tensors.v1"
+@architectures = "spacy.Tok2VecTensors.v1"
 width = ${nlp.pipeline.tok2vec.model:width}
 
 [nlp.pipeline.parser.model]
-@architectures = "transition_based_parser.v1"
+@architectures = "spacy.TransitionBasedParser.v1"
 nr_feature_tokens = 8
 hidden_width = 64
 maxout_pieces = 3
 
 [nlp.pipeline.parser.model.tok2vec]
-@architectures = "tok2vec_tensors.v1"
+@architectures = "spacy.Tok2VecTensors.v1"
 width = ${nlp.pipeline.tok2vec.model:width}
 
 [nlp.pipeline.tok2vec.model]
-@architectures = "hash_embed_cnn.v1"
+@architectures = "spacy.HashEmbedCNN.v1"
 pretrained_vectors = ${nlp:vectors}
 width = 96
 depth = 4
