+++ spacy/ml/models/__init__.py
@@ -1,0 +1,6 @@
+from .morphologizer import build_morphologizer_model
+from .parser import build_old_parser_model
+from .tagger import build_tagger_model
+from .tensorizer import build_tensorizer
+from .tok2vec import build_Tok2Vec_model
+from .multi_task import build_masked_language_model
