+++ spacy/ml/models/tagger_defaults.cfg
@@ -1,0 +1,9 @@
+[model]
+@architectures = "spacy.TaggerModel.v1"
+pretrained_vectors = null
+width = 96
+depth = 4
+embed_size = 2000
+window_size = 1
+maxout_pieces = 3
+subword_features = True
